Shift common utilities to chat template utils.
