+++ src/transformers/utils/chat_template_utils.py
@@ -53,6 +53,9 @@ if is_torch_available():
     from torch import Tensor
 
 
+ChatType = list[dict[str, str]]
+
+
 BASIC_TYPES = (int, float, str, bool, Any, type(None), ...)
 # Extracts the initial segment of the docstring, containing the function description
 description_re = re.compile(r"^(.*?)[\n\s]*(Args:|Returns:|Raises:|\Z)", re.DOTALL)
@@ -459,9 +462,9 @@ def _compile_jinja_template(chat_template):
 
 
 def render_jinja_template(
-    conversations: list[list[dict[str, str]]],
+    conversations: list[ChatType],
     tools: list[dict | Callable] | None = None,
-    documents: list[dict[str, str]] | None = None,
+    documents: ChatType | None = None,
     chat_template: str | None = None,
     return_assistant_tokens_mask: bool = False,
     continue_final_message: bool = False,
@@ -558,3 +561,15 @@ def render_jinja_template(
         rendered.append(rendered_chat)
 
     return rendered, all_generation_indices
+
+
+class Chat:
+    """This class is intended to just be used internally for pipelines and not exposed to users. We convert chats
+    to this format because the rest of the pipeline code tends to assume that lists of messages are
+    actually a batch of samples rather than messages in the same conversation."""
+
+    def __init__(self, messages: dict):
+        for message in messages:
+            if not ("role" in message and "content" in message):
+                raise ValueError("When passing chat dicts as input, each dict must have a 'role' and 'content' key.")
+        self.messages = messages
